+++ src/common.ts
@@ -3,6 +3,10 @@ import { types } from 'vortex-api';
 import { IPluginRequirement } from './types';
 import { findModByFile } from './util';
 
+export const NAMESPACE = 'game-palworld';
+
+export const NOTIF_ID_BP_MODLOADER_DISABLED = 'notif-palworld-bp-modloader-disabled';
+
 export const DEFAULT_EXECUTABLE = 'Palworld.exe'; // path to executable, relative to game root
 export const XBOX_EXECUTABLE = 'gamelaunchhelper.exe';
 
@@ -13,12 +17,13 @@ export const STEAMAPP_ID = '1623730';
 
 export const UE4SS_PATH_PREFIX = path.join('Pal', 'Binaries');
 export const PAK_MODSFOLDER_PATH = path.join('Pal', 'Content', 'Paks', '~mods'); // relative to game root
+export const BPPAK_MODSFOLDER_PATH = path.join('Pal', 'Content', 'Paks', 'LogicMods');
 
 export const LUA_EXTENSIONS = ['.lua'];
 export const PAK_EXTENSIONS = ['.pak', '.utoc', '.ucas'];
 export const IGNORE_CONFLICTS = ['ue4sslogicmod.info', '.ue4sslogicmod', '.logicmod'];
-export const IGNORE_DEPLOY = [path.join('**', 'enabled.txt')];
 
+export const UE4SS_ENABLED_FILE = 'enabled.txt';
 export const XBOX_UE4SS_XINPUT_REPLACEMENT = 'xinput1_4.dll';
 export const UE4SS_FILES = [
   'xinput1_3.dll', 'UE4SS-settings.ini',
@@ -30,6 +35,7 @@ export const TOP_LEVEL_DIRECTORIES = [
 
 export const MOD_TYPE_PAK = 'palworld-pak-modtype';
 export const MOD_TYPE_LUA = 'palworld-lua-modtype';
+export const MOD_TYPE_BP_PAK = 'palworld-blueprint-modtype';
 
 export const UE4SS_XINPUT_FILENAME = 'UE4SS_Xinput_v2.5.2.zip';